move makeRoomForArrows into Arrow.ts and use it there, too
